fix for not doing diff on captured time, larger circular buffer, write removing RX data on TX put

+++ ce_fdd_mfm/Core/Src/circularbuffer.c
@@ -1,7 +1,7 @@
 #include "circularbuffer.h"
 
-uint8_t txCnt, txStore, txLoad;
+uint32_t txCnt, txStore, txLoad;
 uint8_t txData[BFR_SIZE];
 
-uint8_t rxCnt, rxStore, rxLoad;
+uint32_t rxCnt, rxStore, rxLoad;
 uint8_t rxData[BFR_SIZE];
